requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.9.0
Flask-Session==0.8.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-dotenv==1.0.1
gunicorn==23.0.0
python-magic==0.4.27
